feat: add Nix flake
Introduce a Nix flake, a [recently introduced
concept](https://www.tweag.io/blog/2020-05-25-flakes/) in the Nix
community.

This will allow our users and developers to optionally use Nix to work
with `pgx`. This commit is a minimal step, and does not include
documentation or usage guidelines, as it is not officially supported for
users yet.

* Introduce a `cargo-pgx` derivation for Nix users.

  Users can run `cargo-pgx` ad-hoc:

  ```bash
  nix run .#cargo-pgx -- pgx init
  ```

  Or they can develop `cargo-pgx` itself:

  ```bash
  nix develop .#cargo-pgx
  ```
* Add a flake file that imports the `cargo-pgx` derivation.
* Set `cargo-pgx` as the default run/build target:

  ```bash
  nix run .
  nix build .
  ```
* Define a development shell for building `pgx` based extensions.

  ```bash
  nix develop .
  ```

* Define checks for formatting (using `rustfmt` and `nixpkgs-fmt`):

  ```bash
  nix flake check
  ```

* Adds a `.envrc` compatible with
  [`nix-direnv`](https://github.com/nix-community/nix-direnv).

Signed-off-by: Ana Hobden <operator@hoverbear.org>